requirements.txt
pycryptodome==3.19.0
pyperclip==1.8.2